streamlit-invoice-app/requirements.txt
streamlit
bcrypt
plotly
reportlab
arabic_reshaper
python-bidi
pandas